llama-index-tools-maxia/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llama-index-tools-maxia"
version = "0.1.0"
description = "LlamaIndex tools for MAXIA — AI-to-AI Marketplace on 14 blockchains"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "MAXIA", email = "contact@maxiaworld.app" },
]
keywords = [
    "llamaindex",
    "llama-index",
    "maxia",
    "ai-marketplace",
    "solana",
    "crypto",
    "defi",
    "gpu-rental",
    "tokenized-stocks",
    "agent",
    "a2a",
    "tool",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "llama-index-core>=0.11",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "respx>=0.21",
]

[project.urls]
Homepage = "https://maxiaworld.app"
Repository = "https://github.com/majorelalexis-stack/maxia"
Documentation = "https://maxiaworld.app/docs"
"API Reference" = "https://maxiaworld.app/api/public/docs"
"MCP Server" = "https://maxiaworld.app/mcp/manifest"
Issues = "https://github.com/majorelalexis-stack/maxia/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/llama_index_tools_maxia"]
